Keep secret-agent in a single package and make it backwards compatible
