+++ prow/cmd/crier/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
     deps = [
         "//prow/apis/prowjobs/v1:go_default_library",
         "//prow/config:go_default_library",
-        "//prow/config/secret/v2:go_default_library",
+        "//prow/config/secret:go_default_library",
         "//prow/crier:go_default_library",
         "//prow/crier/reporters/gcs:go_default_library",
         "//prow/crier/reporters/gcs/kubernetes:go_default_library",
