+++ prow/cmd/generic-autobumper/bumper/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "k8s.io/test-infra/prow/cmd/generic-autobumper/bumper",
     visibility = ["//visibility:public"],
     deps = [
-        "//prow/config/secret/v2:go_default_library",
+        "//prow/config/secret:go_default_library",
         "//prow/github:go_default_library",
         "//robots/pr-creator/updater:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
@@ -32,5 +32,5 @@ go_test(
     srcs = ["bumper_test.go"],
     data = glob(["testdata/**"]),
     embed = [":go_default_library"],
-    deps = ["//prow/config/secret/v2:go_default_library"],
+    deps = ["//prow/config/secret:go_default_library"],
 )
