+++ prow/cmd/generic-autobumper/bumper/bumper.go
@@ -29,7 +29,7 @@ import (
 
 	"github.com/sirupsen/logrus"
 
-	"k8s.io/test-infra/prow/config/secret/v2"
+	"k8s.io/test-infra/prow/config/secret"
 	"k8s.io/test-infra/prow/github"
 	"k8s.io/test-infra/robots/pr-creator/updater"
 )
