+++ prow/cmd/hook/BUILD.bazel
@@ -46,7 +46,7 @@ go_library(
         "//pkg/flagutil:go_default_library",
         "//prow/bugzilla:go_default_library",
         "//prow/config:go_default_library",
-        "//prow/config/secret/v2:go_default_library",
+        "//prow/config/secret:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/flagutil/config:go_default_library",
         "//prow/flagutil/plugins:go_default_library",
