+++ prow/cmd/invitations-accepter/BUILD.bazel
@@ -18,7 +18,6 @@ go_library(
     importpath = "k8s.io/test-infra/prow/cmd/invitations-accepter",
     visibility = ["//visibility:private"],
     deps = [
-        "//prow/config/secret/v2:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/github:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
