+++ prow/cmd/jenkins-operator/BUILD.bazel
@@ -23,7 +23,7 @@ go_library(
     importpath = "k8s.io/test-infra/prow/cmd/jenkins-operator",
     deps = [
         "//pkg/flagutil:go_default_library",
-        "//prow/config/secret/v2:go_default_library",
+        "//prow/config/secret:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/flagutil/config:go_default_library",
         "//prow/interrupts:go_default_library",
