+++ prow/cmd/tackle/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
     importpath = "k8s.io/test-infra/prow/cmd/tackle",
     visibility = ["//visibility:public"],
     deps = [
-        "//prow/config/secret/v2:go_default_library",
+        "//prow/config/secret:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/github:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
