--- prow/config/secret/v2/BUILD.bazel
@@ -1,43 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "agent.go",
-        "secret.go",
-    ],
-    importpath = "k8s.io/test-infra/prow/config/secret/v2",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//prow/logrusutil:go_default_library",
-        "//prow/secretutil:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-        "@io_k8s_apimachinery//pkg/util/sets:go_default_library",
-    ],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-    visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["agent_test.go"],
-    embed = [":go_default_library"],
-    deps = [
-        "//prow/logrusutil:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-    ],
-)
