+++ prow/external-plugins/needs-rebase/BUILD.bazel
@@ -17,7 +17,7 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//pkg/flagutil:go_default_library",
-        "//prow/config/secret/v2:go_default_library",
+        "//prow/config/secret:go_default_library",
         "//prow/external-plugins/needs-rebase/plugin:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/flagutil/plugins:go_default_library",
