+++ prow/flagutil/BUILD.bazel
@@ -22,7 +22,7 @@ go_library(
         "//prow/bugzilla:go_default_library",
         "//prow/client/clientset/versioned:go_default_library",
         "//prow/client/clientset/versioned/typed/prowjobs/v1:go_default_library",
-        "//prow/config/secret/v2:go_default_library",
+        "//prow/config/secret:go_default_library",
         "//prow/git:go_default_library",
         "//prow/git/v2:go_default_library",
         "//prow/github:go_default_library",
@@ -73,7 +73,6 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//pkg/flagutil:go_default_library",
-        "//prow/config/secret/v2:go_default_library",
         "//prow/github:go_default_library",
         "@com_github_google_go_cmp//cmp:go_default_library",
     ],
