+++ prow/flagutil/jira.go
@@ -22,7 +22,7 @@ import (
 	"fmt"
 	"net/url"
 
-	"k8s.io/test-infra/prow/config/secret/v2"
+	"k8s.io/test-infra/prow/config/secret"
 	"k8s.io/test-infra/prow/jira"
 )
