+++ prow/ghhook/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
     importpath = "k8s.io/test-infra/prow/ghhook",
     visibility = ["//visibility:public"],
     deps = [
-        "//prow/config/secret/v2:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/github:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
