+++ robots/commenter/BUILD.bazel
@@ -31,7 +31,7 @@ go_library(
     srcs = ["main.go"],
     importpath = "k8s.io/test-infra/robots/commenter",
     deps = [
-        "//prow/config/secret/v2:go_default_library",
+        "//prow/config/secret:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/github:go_default_library",
     ],
