+++ robots/pr-creator/BUILD.bazel
@@ -7,7 +7,6 @@ go_library(
     importpath = "k8s.io/test-infra/robots/pr-creator",
     visibility = ["//visibility:private"],
     deps = [
-        "//prow/config/secret/v2:go_default_library",
         "//prow/flagutil:go_default_library",
         "//robots/pr-creator/updater:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
